refactor: split FullOperation::ProgramFactory into separate files
- Extract ProgramFactory implementation into full_program_factory.cpp
- Create shared types in full_common_types.hpp
- Extract utility functions to full_program_factory_common.hpp
- Update CMakeLists.txt to include new files
- Maintain correct API usage and build compatibility
