+++ ttnn/cpp/ttnn/operations/full/device/full_device_operation.cpp
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "full_device_operation.hpp"
+#include "full_program_factory.hpp"
 
 #include "ttnn/tensor/tensor.hpp"
