+++ ttnn/cpp/ttnn/operations/full/device/full_program_factory.hpp
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+#include "full_device_operation.hpp"
